+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Added
 - Made `relations` and `potential_schema` optional in `SchemaBuilder`.
+- Added a check to prevent the use of deprecated Cypher syntax for Neo4j versions 5.23.0 and above.
+
+### Changed
+- Vector and Hybrid retrievers used with `return_properties` now also return the node labels (`nodeLabels`) and the node's element ID (`id`).
+- `HybridRetriever` now filters out the embedding property index in `self.vector_index_name` from the retriever result by default.
+
 
 ## 1.1.0